Unused modules and components removed, auth folder containing guards renamed to 'guards'
