+++ apps/frontend/src/app/guards/auth.guard.ts
