--- apps/frontend/src/app/material/material.module.ts
@@ -1,22 +1,0 @@
-import { NgModule } from '@angular/core';
-import {MatSidenavModule} from '@angular/material/sidenav'
-import {MatIconModule} from '@angular/material/icon'
-import {MatToolbarModule} from '@angular/material/toolbar'
-import {MatListModule} from '@angular/material/list'
-
-const MaterialComponents = [
-  MatSidenavModule,
-  MatIconModule,
-  MatToolbarModule,
-  MatListModule
-]
-
-@NgModule({
-  imports: [
-    MaterialComponents
-  ],
-  exports:[
-    MaterialComponents
-  ]
-})
-export class MaterialModule { }
