--- apps/frontend/src/app/modules/currency-dropdown/currency-dropdown.component.scss
@@ -1,7 +1,0 @@
-mat-form-field {
-    width: 100%;
-  }
-  mat-form-field,
-  input {
-    font-size: 14px;
-  }
