--- apps/frontend/src/app/modules/currency-dropdown/currency-dropdown.component.spec.ts
@@ -1,22 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { CurrencyDropdownComponent } from './currency-dropdown.component';
-
-describe('CurrencyDropdownComponent', () => {
-  let component: CurrencyDropdownComponent;
-  let fixture: ComponentFixture<CurrencyDropdownComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [CurrencyDropdownComponent],
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(CurrencyDropdownComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
