--- apps/frontend/src/app/navigation/navigation.component.html
@@ -1,1 +1,0 @@
-<p>navigation works!</p>
