--- apps/frontend/src/app/navigation/navigation.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'crtvs-navigation',
-  templateUrl: './navigation.component.html',
-  styleUrls: ['./navigation.component.scss'],
-})
-export class NavigationComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
-}
